readd casts

+++ core/src/Cuda/Kokkos_Cuda_Half_MathematicalFunctions.hpp
@@ -126,10 +126,10 @@ KOKKOS_CUDA_HALF_AND_BHALF_UNARY_FUNCTION_IMPL(rint, hrint)
 // as the range of int is enough for any value half_t can take.
 // Thus we just cast to the required return type here
 KOKKOS_INLINE_FUNCTION long impl_lrint(Kokkos::Experimental::half_t x) {
-  return impl_rint(x);
+  return static_cast<long>(impl_rint(x));
 }
 KOKKOS_INLINE_FUNCTION long long impl_llrint(Kokkos::Experimental::half_t x) {
-  return impl_rint(x);
+  return static_cast<long long>(impl_rint(x));
 }
 // logb
 // nextafter
